Switch to xlsx-populate library
