+++ src/data/custom-templates/WMR_IR_Discriminating_Concentration_Bioassay.ts
@@ -1,6 +1,6 @@
-import { ProgramExcelJSTemplate } from "../ExcelTemplate";
+import { XLSXPopulateTemplate } from "../XLSXPopulateTemplate";
 
-export default class extends ProgramExcelJSTemplate {
+export default class extends XLSXPopulateTemplate {
     public readonly id = "WMR_IR_v0";
     public readonly name = "WMR_IR_Discriminating_Concentration_Bioassay";
     public readonly url = "templates/WMR_IR_discriminating_concentration_bioassays_v0_EN.xlsx";
