+++ src/domain/repositories/TemplateProvider.ts
@@ -3,6 +3,6 @@ import { Template } from "../entities/Template";
 
 export interface TemplateProvider {
     templates: Template[];
-    listTemplates(): Pick<Template, "id" | "name" | "type">[];
+    listTemplates(): Pick<Template, "id" | "name">[];
     getTemplate(id: Id): Promise<Template>;
 }
